refactor(Judge): Game 클래스의 일부 기능을 Judge 클래스로 분리
- 게임 진행 중 사용자의 개별 입력에 대한 처리만 담당
- Game 클래스 내부에 Judge 인스턴스 포함
- Game과 Judge 모두 인스턴스 1개만 생성하여 관리
